requirements.txt
pandas
numpy
scipy
matplotlib
statsmodels
openpyxl
fpdf2
tqdm